// cypress/support/enums/nav-tabs-enum.ts
export enum PIM_NAV_TABS{
    CONFIGURATION = 'Configuration',
    EMPLOYEE_LIST = 'Employee List',
    ADD_EMPLOYEE = 'Add Employee',
    REPORTS = 'Reports' 
}

export enum LEAVE_NAV_TABS{
    ENTITLEMENTS = 'Entitlements',
    ADD_ENTITLEMENTS = 'Add Entitlements',
    APPLY = 'Apply',
    MY_LEAVE = 'My Leave',
    LEAVE_LIST = 'Leave List'
}

export enum MY_INFO_TABS{
    PERSONAL_DETAILS = 'Personal Details',
    CONTACT_DETAILS = 'Contact Details',
    EMERGENCY_CONTACTS = 'Emergency Contacts',
    DEPENDENTS = 'Dependents',
    IMMIGRATION = 'Immigration',
    JOB = 'Job',
    SALARY = 'Salary',
    REPORT_TO = 'Report-to',
    QUALIFICATIONS = 'Qualifications',
    MEMBERSHIPS = 'Memberships'
}

export enum RECRUITMENT_NAV_TABS{
    CANDIDATES = 'Candidates',
    VACANCIES = 'Vacancies',
}